chore: port AppMap webview to new abstract base classes
